fix: pose-trace at_rest logged as int (was garbage)
The 'v' trace printed at_rest as a bool with %d; Zephyr cbprintf packed
it as 1 byte but unpacked 4, yielding garbage (values ending ...0001 =
real low bit + 3 junk bytes).  Cast to (int) so the stillness flag logs
clean 0/1.  Needed because the stillness gate is now load-bearing in the
observability-aware dictation classifier and must be observable in the
adversarial measurement traces.

+++ src/gesture_mode.cpp
@@ -760,7 +760,7 @@ static void pose_trace_tick(void)
             "pitch=%.0f roll=%.0f at_rest=%d",
             (double)gx_filt, (double)gy_filt, (double)gz_filt,
             (double)shadow, (double)vert,
-            (double)ori.pitch_deg, (double)ori.roll_deg, ori.at_rest);
+            (double)ori.pitch_deg, (double)ori.roll_deg, (int)ori.at_rest);
 }
 
 void gesture_mode_update_accel(float ax, float ay, float az)
